Rename sentry module
